python_duckdb/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
anyio==4.9.0
    # via httpx
awscrt==0.23.8
    # via botocore
black==25.1.0
    # via -r .\requirements.in
boto3[crt]==1.37.21
    # via -r .\requirements.in
botocore[crt]==1.37.21
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via black
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage[toml]==7.7.1
    # via pytest-cov
docopt==0.6.2
    # via -r .\requirements.in
duckdb==1.2.1
    # via -r .\requirements.in
faker==37.1.0
    # via -r .\requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r .\requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via -r .\requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==3.0.2
    # via jinja2
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.7
    # via black
pluggy==1.5.0
    # via pytest
psutil==7.0.0
    # via -r .\requirements.in
pytest==8.3.5
    # via
    #   -r .\requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.26.0
    # via -r .\requirements.in
pytest-cov==6.0.0
    # via -r .\requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.0
    # via -r .\requirements.in
pytz==2025.2
    # via -r .\requirements.in
redis==5.2.1
    # via -r .\requirements.in
s3transfer==0.11.4
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
typing-extensions==4.13.0
    # via anyio
tzdata==2025.2
    # via faker
urllib3==2.3.0
    # via botocore
